Task 20 correct average score data type

+++ pipeline/databases/20-average_score.sql
@@ -3,7 +3,7 @@ DELIMITER //
 
 CREATE PROCEDURE ComputeAverageScoreForUser (IN user_id_p INT)
 BEGIN
-    DECLARE avg_score INT;
+    DECLARE avg_score FLOAT;
     
     SELECT AVG(score) INTO avg_score FROM corrections WHERE user_id = user_id_p;
